Added Conditional Generative Adversarial Network implementation with Generator, Discriminator, and training loop for image generation.
A Conditional Generative Adversarial Network (cGAN) is an extension of the standard GAN where the generation of data is conditioned on additional information, such as class labels or attributes. It allows for greater control over the output by guiding the Generator to produce specific types of data based on the given condition.